#rb none
#rnx

[CL 14967557 by Ben Marsh in ue5-main branch]

+++ Engine/Source/Programs/Shared/EpicGames.Core/FileUtils.cs
@@ -1,11 +1,14 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
+using Microsoft.Win32.SafeHandles;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.ComTypes;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -46,6 +49,11 @@ namespace EpicGames.Core
 		public static IEqualityComparer<string> PlatformPathComparer = Environment.OSVersion.Platform == PlatformID.Unix ? StringComparer.Ordinal : StringComparer.OrdinalIgnoreCase;
 
 		/// <summary>
+		/// Utf8 string comparer that should be used for native path comparisons
+		/// </summary>
+		public static IEqualityComparer<ReadOnlyUtf8String> PlatformPathComparerUtf8 = Environment.OSVersion.Platform == PlatformID.Unix ? ReadOnlyUtf8StringComparer.Ordinal : ReadOnlyUtf8StringComparer.OrdinalIgnoreCase;
+
+		/// <summary>
 		/// Finds the on-disk case of a a file
 		/// </summary>
 		/// <param name="Info">FileInfo instance describing the file</param>
@@ -71,7 +79,11 @@ namespace EpicGames.Core
 		{
 			if (!DirectoryReference.Exists(Directory))
 			{
-				CreateDirectoryTree(Directory.ParentDirectory);
+				DirectoryReference? ParentDirectory = Directory.ParentDirectory;
+				if (ParentDirectory != null)
+				{
+					CreateDirectoryTree(ParentDirectory);
+				}
 				DirectoryReference.CreateDirectory(Directory);
 			}
 		}
@@ -82,7 +94,14 @@ namespace EpicGames.Core
 		/// <param name="FileName">Name of the file to delete</param>
 		public static void ForceDeleteFile(string FileName)
 		{
-			ForceDeleteFile(new FileInfo(FileName));
+			if (Environment.OSVersion.Platform == PlatformID.Win32NT)
+			{
+				ForceDeleteFileWin32(FileName);
+			}
+			else
+			{
+				ForceDeleteFile(new FileInfo(FileName));
+			}
 		}
 
 		/// <summary>
@@ -91,17 +110,25 @@ namespace EpicGames.Core
 		/// <param name="File">The file to delete</param>
 		public static void ForceDeleteFile(FileInfo File)
 		{
-			try
+			if (Environment.OSVersion.Platform == PlatformID.Win32NT)
 			{
-				if (File.Exists)
-				{
-					File.Attributes = FileAttributes.Normal;
-					File.Delete();
-				}
+				ForceDeleteFileWin32(File.FullName);
+				File.Refresh();
 			}
-			catch (Exception Ex)
+			else
 			{
-				throw new WrappedFileOrDirectoryException(Ex, String.Format("Unable to delete '{0}': {1}", File.FullName, Ex.Message.TrimEnd()));
+				try
+				{
+					if (File.Exists)
+					{
+						File.Attributes = FileAttributes.Normal;
+						File.Delete();
+					}
+				}
+				catch (Exception Ex)
+				{
+					throw new WrappedFileOrDirectoryException(Ex, String.Format("Unable to delete '{0}': {1}", File.FullName, Ex.Message.TrimEnd()));
+				}
 			}
 		}
 
@@ -111,7 +138,14 @@ namespace EpicGames.Core
 		/// <param name="Location">The file to delete</param>
 		public static void ForceDeleteFile(FileReference Location)
 		{
-			ForceDeleteFile(new FileInfo(Location.FullName));
+			if (Environment.OSVersion.Platform == PlatformID.Win32NT)
+			{
+				ForceDeleteFileWin32(Location.FullName);
+			}
+			else
+			{
+				ForceDeleteFile(new FileInfo(Location.FullName));
+			}
 		}
 
 		/// <summary>
@@ -196,19 +230,22 @@ namespace EpicGames.Core
 			}
 			else
 			{
-				foreach (FileInfo File in Directory.EnumerateFiles())
+				if (Directory.Exists)
 				{
-					ForceDeleteFile(File);
-				}
-				foreach (DirectoryInfo SubDirectory in Directory.EnumerateDirectories())
-				{
-					if (SubDirectory.Attributes.HasFlag(FileAttributes.ReparsePoint))
+					foreach (FileInfo File in Directory.EnumerateFiles())
 					{
-						ForceDeleteDirectoryInternal(SubDirectory);
+						ForceDeleteFile(File);
 					}
-					else
+					foreach (DirectoryInfo SubDirectory in Directory.EnumerateDirectories())
 					{
-						ForceDeleteDirectory(SubDirectory);
+						if (SubDirectory.Attributes.HasFlag(FileAttributes.ReparsePoint))
+						{
+							ForceDeleteDirectoryInternal(SubDirectory);
+						}
+						else
+						{
+							ForceDeleteDirectory(SubDirectory);
+						}
 					}
 				}
 			}
@@ -277,7 +314,7 @@ namespace EpicGames.Core
 				}
 
 				// Throw the original exception
-				throw new WrappedFileOrDirectoryException(Ex, String.Format("{0}: Unable to move {1} to {2}", Ex.GetType(), SourceLocation, TargetLocation));
+				throw new WrappedFileOrDirectoryException(Ex, String.Format("Unable to move {0} to {1}", SourceLocation, TargetLocation));
 			}
 		}
 
@@ -329,6 +366,8 @@ namespace EpicGames.Core
 
 		static readonly IntPtr INVALID_HANDLE_VALUE = new IntPtr(-1);
 
+		const int ERROR_PATH_NOT_FOUND = 3;
+
 		private static void ForceDeleteLongDirectoryContentsWin32(string DirName)
 		{
 			WIN32_FIND_DATA FindData = new WIN32_FIND_DATA();
@@ -354,10 +393,7 @@ namespace EpicGames.Core
 							{
 								SetFileAttributesW(FullName, FILE_ATTRIBUTE_NORMAL);
 							}
-							if (!DeleteFileW(FullName))
-							{
-								throw new WrappedFileOrDirectoryException(new Win32Exception(), "Unable to delete " + FullName);
-							}
+							ForceDeleteFileWin32(FullName);
 						}
 
 						if (!FindNextFileW(hFind, ref FindData))
@@ -379,8 +415,280 @@ namespace EpicGames.Core
 
 			if (!RemoveDirectory(DirName))
 			{
-				throw new WrappedFileOrDirectoryException(new Win32Exception(), "Unable to delete " + DirName);
+				int ErrorCode = Marshal.GetLastWin32Error();
+				if (ErrorCode != ERROR_PATH_NOT_FOUND)
+				{
+					throw new WrappedFileOrDirectoryException(new Win32Exception(ErrorCode), "Unable to delete " + DirName);
+				}
 			}
+		}
+
+		const int ERROR_FILE_NOT_FOUND = 2;
+		const int ERROR_ACCESS_DENIED = 5;
+
+		[DllImport("kernel32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
+		internal static extern int GetFileAttributesW(string lpFileName);
+
+		public static void ForceDeleteFileWin32(string FileName)
+		{
+			// Try to delete the file normally
+			if (DeleteFileW(FileName))
+			{
+				return;
+			}
+
+			// Capture the exception for failing to delete the file
+			Win32Exception Ex = new Win32Exception();
+
+			// Check the file exists and is not readonly
+			int Attributes = GetFileAttributesW(FileName);
+			if (Attributes == -1)
+			{
+				int ErrorCode = Marshal.GetLastWin32Error();
+				if (ErrorCode == ERROR_PATH_NOT_FOUND || ErrorCode == ERROR_FILE_NOT_FOUND)
+				{
+					return;
+				}
+			}
+			else
+			{
+				if ((Attributes & (int)FileAttributes.ReadOnly) != 0)
+				{
+					if (SetFileAttributesW(FileName, (int)FileAttributes.Normal) && DeleteFileW(FileName))
+					{
+						return;
+					}
+				}
+			}
+
+			// Get a useful error message about why the delete failed
+			StringBuilder Message = new StringBuilder($"Unable to delete {FileName} - {Ex.Message}");
+			if (Ex.NativeErrorCode == ERROR_ACCESS_DENIED)
+			{
+				List<FileLockInfo_Win32>? LockInfoList;
+				try
+				{
+					LockInfoList = GetFileLockInfo_Win32(FileName);
+				}
+				catch
+				{
+					LockInfoList = null;
+				}
+
+				if (LockInfoList != null && LockInfoList.Count > 0)
+				{
+					Message.Append("\nProcesses with open handles to file:");
+					foreach (FileLockInfo_Win32 LockInfo in LockInfoList)
+					{
+						Message.Append($"\n  {LockInfo}");
+					}
+				}
+			}
+			throw new WrappedFileOrDirectoryException(Ex, Message.ToString());
+		}
+
+		#endregion
+
+		#region Win32 Restart Manager API
+
+		[StructLayout(LayoutKind.Sequential)]
+		struct RM_UNIQUE_PROCESS
+		{
+			public int dwProcessId;
+			public FILETIME ProcessStartTime;
+		}
+
+		const int RmRebootReasonNone = 0;
+		static readonly int RM_SESSION_KEY_LEN = Marshal.SizeOf<Guid>();
+		static readonly int CCH_RM_SESSION_KEY = RM_SESSION_KEY_LEN * 2;
+		const int CCH_RM_MAX_APP_NAME = 255;
+		const int CCH_RM_MAX_SVC_NAME = 63;
+
+		enum RM_APP_TYPE
+		{
+			RmUnknownApp = 0,
+			RmMainWindow = 1,
+			RmOtherWindow = 2,
+			RmService = 3,
+			RmExplorer = 4,
+			RmConsole = 5,
+			RmCritical = 1000
+		}
+
+		[StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+		struct RM_PROCESS_INFO
+		{
+			public RM_UNIQUE_PROCESS Process;
+
+			[MarshalAs(UnmanagedType.ByValTStr, SizeConst = CCH_RM_MAX_APP_NAME + 1)]
+			public string strAppName;
+
+			[MarshalAs(UnmanagedType.ByValTStr, SizeConst = CCH_RM_MAX_SVC_NAME + 1)]
+			public string strServiceShortName;
+
+			public RM_APP_TYPE ApplicationType;
+			public uint AppStatus;
+			public uint TSSessionId;
+
+			[MarshalAs(UnmanagedType.Bool)]
+			public bool bRestartable;
+		}
+
+		[DllImport("rstrtmgr.dll", CharSet = CharSet.Unicode)]
+		static extern int RmRegisterResources(uint pSessionHandle, uint nFiles, string[] rgsFilenames, uint nApplications, [In] RM_UNIQUE_PROCESS[]? rgApplications, uint nServices, string[]? rgsServiceNames);
+
+		[DllImport("rstrtmgr.dll", CharSet = CharSet.Auto)]
+		static extern int RmStartSession(out uint pSessionHandle, int dwSessionFlags, StringBuilder strSessionKey);
+
+		[DllImport("rstrtmgr.dll")]
+		static extern int RmEndSession(uint pSessionHandle);
+
+		[DllImport("rstrtmgr.dll")]
+		static extern int RmGetList(uint dwSessionHandle,
+									out uint pnProcInfoNeeded,
+									ref uint pnProcInfo,
+									[In, Out] RM_PROCESS_INFO[] rgAffectedApps,
+									ref uint lpdwRebootReasons);
+
+		const int PROCESS_QUERY_LIMITED_INFORMATION = 0x1000;
+
+		[DllImport("kernel32.dll", SetLastError = true)]
+		static extern SafeProcessHandle OpenProcess(
+			 int processAccess,
+			 bool bInheritHandle,
+			 int processId
+		);
+
+		[DllImport("kernel32.dll", SetLastError = true)]
+		[return: MarshalAs(UnmanagedType.Bool)]
+		static extern bool GetProcessTimes(SafeProcessHandle hProcess, out FILETIME
+		   lpCreationTime, out FILETIME lpExitTime, out FILETIME lpKernelTime,
+		   out FILETIME lpUserTime);
+
+		[DllImport("kernel32.dll", SetLastError = true)]
+		static extern bool QueryFullProcessImageName([In]SafeProcessHandle hProcess, [In]int dwFlags, [Out]StringBuilder lpExeName, ref int lpdwSize);
+
+		/// <summary>
+		/// Information about a locked file
+		/// </summary>
+		public class FileLockInfo_Win32
+		{
+			/// <summary>
+			/// Process id
+			/// </summary>
+			public int ProcessId;
+
+			/// <summary>
+			/// Path to the process holding the lock
+			/// </summary>
+			public string? FileName;
+
+			/// <summary>
+			/// Name of the application
+			/// </summary>
+			public string AppName;
+
+			/// <summary>
+			/// Time at which the process started
+			/// </summary>
+			public DateTime StartTime;
+
+			/// <summary>
+			/// Constructor
+			/// </summary>
+			/// <param name="ProcessId">Process id</param>
+			/// <param name="FileName">Path to the process holding the lock</param>
+			/// <param name="AppName">Name of the application</param>
+			/// <param name="StartTime">Time at which the process started</param>
+			public FileLockInfo_Win32(int ProcessId, string? FileName, string AppName, DateTime StartTime)
+			{
+				this.ProcessId = ProcessId;
+				this.FileName = FileName;
+				this.AppName = AppName;
+				this.StartTime = StartTime;
+			}
+
+			/// <inheritdoc/>
+			public override string ToString()
+			{
+				return $"{ProcessId}: {FileName ?? AppName} (started {StartTime})";
+			}
+		}
+
+		/// <summary>
+		/// Gets a list of processes that have a handle to the given file open
+		/// </summary>
+		/// <param name="FileName">File to check</param>
+		/// <returns>List of processes with a lock open</returns>
+		public static List<FileLockInfo_Win32> GetFileLockInfo_Win32(string FileName)
+		{
+			uint SessionHandle = 0;
+			try
+			{
+				StringBuilder SessionKey = new StringBuilder(CCH_RM_SESSION_KEY + 1);
+
+				int Result = RmStartSession(out SessionHandle, 0, SessionKey);
+				if (Result != 0)
+				{
+					throw new Win32Exception(Result, "Unable to open restart manager session");
+				}
+
+				Result = RmRegisterResources(SessionHandle, 1, new string[] { FileName }, 0, null, 0, null);
+				if (Result != 0)
+				{
+					throw new Win32Exception(Result, "Unable to register resource with restart manager");
+				}
+
+				uint nProcInfoNeeded = 0;
+				uint nProcInfo = 10;
+				uint Reason = 0;
+				RM_PROCESS_INFO[] ProcessInfoArray = new RM_PROCESS_INFO[nProcInfo];
+				Result = RmGetList(SessionHandle, out nProcInfoNeeded, ref nProcInfo, ProcessInfoArray, ref Reason);
+				if (Result != 0)
+				{
+					throw new Win32Exception(Result, "Unable to query processes with file handle open");
+				}
+
+				List<FileLockInfo_Win32> FileLocks = new List<FileLockInfo_Win32>();
+				for (int Idx = 0; Idx < nProcInfo; Idx++)
+				{
+					RM_PROCESS_INFO ProcessInfo = ProcessInfoArray[Idx];
+					long StartTimeTicks = FileTimeToTicks(ProcessInfo.Process.ProcessStartTime);
+
+					string? ImageName = null;
+					using (SafeProcessHandle hProcess = OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, false, ProcessInfo.Process.dwProcessId))
+					{
+						if (hProcess != null)
+						{
+							FILETIME CreateTime, ExitTime, KernelTime, UserTime;
+							if (GetProcessTimes(hProcess, out CreateTime, out ExitTime, out KernelTime, out UserTime) && FileTimeToTicks(CreateTime) == StartTimeTicks)
+							{
+								int Capacity = 260;
+								StringBuilder ImageNameBuilder = new StringBuilder(Capacity);
+								if (QueryFullProcessImageName(hProcess, 0, ImageNameBuilder, ref Capacity))
+								{
+									ImageName = ImageNameBuilder.ToString(0, Capacity);
+								}
+							}
+						}
+					}
+
+					FileLocks.Add(new FileLockInfo_Win32(ProcessInfo.Process.dwProcessId, ImageName, ProcessInfo.strAppName, DateTime.FromFileTime(StartTimeTicks)));
+				}
+				return FileLocks;
+			}
+			finally
+			{
+				if (SessionHandle != 0)
+				{
+					RmEndSession(SessionHandle);
+				}
+			}
+		}
+
+		private static long FileTimeToTicks(FILETIME FileTime)
+		{
+			return (long)(uint)FileTime.dwLowDateTime | ((long)(uint)FileTime.dwHighDateTime << 32);
 		}
 
 		#endregion
